The scrollbar is breaking the column builder.
The scrollbar can cause the page to be redrawn in the
middle of the column-building routine, which -may-
throw off the column-height calculations and result
in other errors. This forces the page to be of a
certain height and thus the scrollbar is permanently
visible. There may be a better solution to come.